Add files via upload
 Chatbot Eliza
Objectives
To gain experience with the Python programming language, in particular the use of regular expressions 
in order to do "word spotting" and transform sentences from answers into questions. To gain an 
appreciation for the difficulties of natural language processing in general.
 Specification
Write an Eliza program in Python. No chat packages/functions/libraries are allowed to use. The 
program should be called eliza.py, and it should run from the command line with no arguments. 
Your program should engage in a dialogue with the user, with your program Eliza playing the role of a 
psychotherapist. Your program should be able carry out "word spotting", that is it should recognize 
certain key words and respond simply based on that word being present in the input. It should also be 
able to transform certain simple sentence forms from statements (from the user) into questions (that 
Eliza will ask).